requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
cloudpickle==3.1.1
decorator==5.2.1
distrax==0.1.5
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.28
docker-pycreds==0.4.0
etils==1.12.2
Farama-Notifications==0.0.4
flax==0.10.5
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
humanize==4.12.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
jax==0.6.0
jax-cuda12-pjrt==0.6.0
jax-cuda12-plugin==0.6.0
jaxlib==0.6.0
labmaze==1.0.6
lxml==5.3.2
markdown-it-py==3.0.0
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.1
msgpack==1.1.0
mujoco==3.3.1
nest-asyncio==1.6.0
numpy<2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.8.0.87
nvidia-cufft-cu12==11.3.3.83
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-nccl-cu12==2.26.2.post1
nvidia-nvjitlink-cu12==12.8.93
ogbench==1.1.2
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.12
pillow==11.2.1
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scipy==1.15.2
sentry-sdk==2.26.1
setproctitle==1.3.5
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
tensorflow-probability==0.25.0
tensorstore==0.1.73
toolz==1.0.0
tqdm==4.67.1
treescope==0.1.9
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wandb==0.19.9
wrapt==1.17.2
zipp==3.21.0
h5py
robomimic
wandb[media]
opencv-python
seaborn
scikit-learn
flaxmodels